changed 'ThermoState' to 'StreamThermoState' for more specificity
I made the change as I to make sure that it is understood that the
ThermoState struct is used by the Streams within the flow diagram and
not by the blocks. The blocks (reactors, separators, etc) will contain
their own calculations.

On the other hand, the ThermoPackage may be used by both streams and
blocks